Relocate mongo `line` spec
